dcv: Adapt to multiple sessions
There are going to be multiple GNOME Shell instances running that
will listen to the UserAuthenticated signal. Pass the DCV session
ID as a parameter to allow each instance to detect whether it should
react to that signal. To detect this, each GNOME Shell extension
should be aware of which DCV session it is running in. Take this into
account.